Standardize the menu for new file and auto layout
